Extract src/sidebarview.ts — the sidebar's rows
Third render-module split, same *view.ts boundary: data in, string out, no DOM.
Moved the session/external/dormant row builders, the worktree cluster bodies
that wrap them, the GLYPH/GCLASS vocabulary and taskStateText — 124 lines.
renderSidebar, renderMini and initProjectDnD stay in main.ts: they own the
element, the drag state and the delegated handlers these rows' data- attributes
are read by.

The stage pointer moved to ./state to make it possible. `mirror` (plus
extMirrorId / extMirrorPid / pastMirrorId) is what dormantRow and extRow consult
to mark the mirrored row, and a view module may not import main.ts. It belongs
there anyway: state.ts already holds activeId, and CLAUDE.md's rule is that the
two are mutually exclusive — one stage, one owner. 23 reads stay bare
identifiers on the live binding; the 4 writes go through setMirror.

The moved code is byte-identical to HEAD:src/main.ts apart from `export`
keywords and the order of three declarations (the glyph maps now sit above
extWorking rather than 100 lines below it).

Verified in the browser: a hand-built ProjGroup with three sessions, an external
and a dormant renders 5 rows and 35 nodes with the worktree cluster headers, the
attention glyph and class, the task row's "▶ test" / "exit 2", the external's
jump button, the dormant title and the context percentage all present.

Note for the next slice: driving these through projectList() in the browser does
not work — vite serves an edited module under a ?t= URL, so a dynamic import gets
a second module instance with its own empty state. Build the input by hand.
