+++ src/main.ts
@@ -32,15 +32,18 @@ import {
 } from "./phase";
 import { bumpFrec, frecScore, parsePal, scoreItem, type PalItem } from "./palette";
 import {
-  accentFor, activeId, colorOverrides, dormants, externals, FAVORITES,
-  saveFavorites, saveProjOrder, sessions, setActiveId, setDormants, setExternals,
-  setFavorites, setProjOrder, setSortMode, setWtGroup as setWtGroupState, sortMode,
-  SORT_MODES, wtGroup, type SortMode, type WtGroup,
+  accentFor, activeId, colorOverrides, dormants, externals, extMirrorId,
+  extMirrorPid, FAVORITES, mirror, pastMirrorId, saveFavorites, saveProjOrder,
+  sessions, setActiveId, setDormants, setExternals, setFavorites, setMirror,
+  setProjOrder, setSortMode, setWtGroup as setWtGroupState, sortMode, SORT_MODES,
+  wtGroup, type SortMode, type WtGroup,
 } from "./state";
 import {
-  allProjects, clusterByWorktree, nextAfterClose, orderedSessions, projectList,
-  urgencyRank, type ProjGroup, type WtCluster,
+  allProjects, nextAfterClose, orderedSessions, projectList, urgencyRank,
 } from "./grouping";
+import {
+  dormantBusy, dormantRows, extWorking, GCLASS, GLYPH, groupBody, taskStateText,
+} from "./sidebarview";
 import {
   applyInputs, applyRunner, execCmd, exitWaiters, lastRunnableById, rememberedInput,
   rememberInput, launchWithDeps, RUNNERS, runnerFor, setTaskLauncher, setTaskLogger,
@@ -256,23 +259,9 @@ function queueRosterSave() {
   rosterTimer = window.setTimeout(flushRoster, 1500);
 }
 function flushRoster() { clearTimeout(rosterTimer); rosterSavedAt = Date.now(); saveRoster(); }
-// The stage shows exactly ONE thing: a live Episko session (activeId), a live
-// external session mirrored read-only, or a dormant session restorable from a past
-// run. Holding the two read-only kinds in a single discriminated pointer — rather
-// than a flag per kind — is what stops them fighting over the stage on the next
-// render tick (see the note in renderAll).
-//
-// The "ext" kind also carries the session's `pid`, because its `id` is Claude's
-// runtime session_id and that ROTATES on /clear, /compact and /resume. The pid is
-// what stays stable, so refreshExternals re-binds through it instead of dropping
-// the selection (which used to silently jump the sidebar to an unrelated session).
-// Same rule as Sess.resumeId and the telemetry path: hold the stable handle.
-let mirror: { kind: "ext"; id: string; pid: number } | { kind: "past"; id: string } | null = null;
-const extMirrorId = (): string | null => (mirror?.kind === "ext" ? mirror.id : null);
-const extMirrorPid = (): number | null => (mirror?.kind === "ext" ? mirror.pid : null);
-const pastMirrorId = (): string | null => (mirror?.kind === "past" ? mirror.id : null);
+// The stage pointer (mirror / extMirrorId / pastMirrorId) now lives in ./state
+// beside activeId — the two are mutually exclusive, so they belong together.
 let extTranscriptTimer: number | undefined;
-const extWorking = (e: ExtSession) => !!e.status && !["idle", "sleeping", "done", ""].includes(e.status);
 
 // Uncommitted git state keyed by folder (a session's workdir or an external's cwd),
 // polled by refreshDirtyStates. It's the single source of truth for the sidebar's
@@ -372,8 +361,6 @@ function cleanTitle(t: string, s: Sess): string {
   return x;
 }
 
-const GLYPH: Record<string, string> = { attention: "◆", working: "●", thinking: "●", done: "✓", idle: "○", error: "✕", ended: "·" };
-const GCLASS: Record<string, string> = { attention: "g-attn", working: "g-work", thinking: "g-work", done: "g-done", idle: "g-idle", error: "g-error", ended: "g-ended" };
 const PILL_TEXT: Record<Phase, string> = { idle: "idle", thinking: "thinking…", working: "working…", done: "your turn", error: "error", ended: "ended" };
 
 // ---------- launch ----------
@@ -590,7 +577,7 @@ async function refreshExternals() {
       const e = externals.find((x) => x.session_id === extMirrorId())
         ?? (pid != null ? externals.find((x) => x.pid === pid) : undefined);
       if (e) {
-        mirror = { kind: "ext", id: e.session_id, pid: e.pid };
+        setMirror({ kind: "ext", id: e.session_id, pid: e.pid });
         renderExtHeader(e); renderExtInspector(e);
       } else {
         // Truly gone — fall back to an Episko session or the empty state.
@@ -626,7 +613,7 @@ async function refreshDirtyStates() {
 function openExternal(sid: string) {
   const e = externals.find((x) => x.session_id === sid);
   if (!e) return;
-  mirror = { kind: "ext", id: sid, pid: e.pid };
+  setMirror({ kind: "ext", id: sid, pid: e.pid });
   setActiveId(null);
   for (const x of sessions.values()) x.pane.classList.remove("active");
   ($("empty") as HTMLElement).style.display = "none";
@@ -644,7 +631,7 @@ function openExternal(sid: string) {
 }
 function closeExternalView() {
   if (mirror == null) return;
-  mirror = null;   // clears the ext pid with it — one pointer, one lifetime
+  setMirror(null);   // clears the ext pid with it — one pointer, one lifetime
   clearInterval(extTranscriptTimer);
   ($("extPane") as HTMLElement).hidden = true;
 }
@@ -655,7 +642,7 @@ function closeExternalView() {
 function openDormant(id: string) {
   const d = dormants.find((x) => x.id === id);
   if (!d) return;
-  mirror = { kind: "past", id };
+  setMirror({ kind: "past", id });
   setActiveId(null);
   for (const x of sessions.values()) x.pane.classList.remove("active");
   ($("empty") as HTMLElement).style.display = "none";
@@ -824,94 +811,8 @@ function renderExtInspector(e: ExtSession) {
 // The project/worktree grouping and the sidebar ordering now live in ./grouping;
 // what follows is only the painting of what it returns.
 
-// A stable colour per branch, from the same hash as project accents so the sidebar's
-// colour language stays consistent (a branch and a project just seed different hues).
-const branchHue = (c: WtCluster) => accentFor(c.branch || c.key);
-
-// `chip` (chip mode only) tags the row with its worktree's colour-coded branch,
-// which expands from a bare ⑃ to the branch name on row hover.
-function sessionRow(s: Sess, chip?: WtCluster): string {
-  const k = statusKey(s);
-  // Prefer the abbreviated title; fall back to the branch/worktree only until
-  // Claude sets a title, so idle rows stay identifiable. (Branch is kept in the
-  // stage header — dropped here to save sidebar space.)
-  const label = s.kind === "task" ? `▶ ${s.run?.label ?? "task"}` : s.title || (s.worktree ? `⑃ ${s.branch}` : (s.branch || "session"));
-  // shells have no telemetry phase — show a terminal prompt glyph (a dot once exited).
-  // tasks keep the status glyphs: an exit code *is* a done/error phase, so a red
-  // build reads exactly like a broken session in the rail.
-  const glyph = s.kind === "shell" ? (s.phase === "ended" ? GLYPH.ended : "❯") : GLYPH[k];
-  const gcls = s.kind === "shell" ? (s.phase === "ended" ? GCLASS.ended : "g-idle") : GCLASS[k];
-  const chipHtml = chip
-    ? `<span class="chip" style="--wtc:${branchHue(chip)}"><span class="fork">⑃</span><span class="lbl">${esc(chip.branch)}</span></span>`
-    : "";
-  return `<div class="srow${chip ? " o3" : ""} ${s.id === activeId ? "active" : ""}" data-sel="${s.id}">
-    <span class="sglyph ${gcls}">${glyph}</span>
-    <span class="sbranch" title="${esc(label)}">${esc(label)}</span>${chipHtml}
-    <span class="sctx">${s.kind === "task" ? esc(taskStateText(s)) : s.ctxPct != null ? Math.round(s.ctxPct) + "%" : ""}</span>
-    <span class="sclose" data-close="${s.id}" title="Close session">✕</span></div>`;
-}
-// The full body of a project group (owned sessions + external rows), shaped by the
-// worktree-grouping mode. subheader → ⑃ cluster headers with nested rows; chip →
-// flat rows each tagged with a colour-coded branch chip; off/toplevel → plain flat
-// rows. A single-checkout project (one cluster) always renders flat — nothing to
-// disambiguate. Externals cluster right alongside owned sessions (same checkout dir).
-function groupBody(p: ProjGroup): string {
-  const flat = () => p.sessions.map((s) => sessionRow(s)).join("") + p.externals.map((e) => extRow(e)).join("");
-  if (wtGroup === "subheader") {
-    const cl = clusterByWorktree(p);
-    if (cl.length >= 2) return cl.map((c) => {
-      const col = branchHue(c), n = c.sessions.length + c.externals.length;
-      const body = c.sessions.map((s) => sessionRow(s)).join("") + c.externals.map((e) => extRow(e)).join("");
-      return `<div class="wthead"><span class="wtglyph" style="color:${col}">⑃</span>`
-        + `<span class="wtname" style="color:${col}" title="${esc(c.branch)}">${esc(c.branch)}</span>`
-        + `<span class="wtcount">${n}</span></div>`
-        + `<div class="wtsessions" style="--wtc:${col}">${body}</div>`;
-    }).join("");
-  } else if (wtGroup === "chip") {
-    const cl = clusterByWorktree(p);
-    if (cl.length >= 2) {
-      const byKey = new Map(cl.map((c) => [c.key, c]));
-      return p.sessions.map((s) => sessionRow(s, byKey.get(s.workdir || p.path))).join("")
-        + p.externals.map((e) => extRow(e, byKey.get(e.cwd || p.path))).join("");
-    }
-  }
-  return flat();
-}
-// Dormant rows always sit below the live ones, outside any worktree cluster.
-function dormantRows(p: ProjGroup): string {
-  return p.dormants.map((d) => dormantRow(d)).join("");
-}
-function dormantRow(d: Restorable): string {
-  const busy = dormantBusy(d);
-  const label = d.title || (d.worktree ? `⑃ ${d.branch}` : d.branch) || "session";
-  const when = relTime(d.lastActivity);
-  const tip = busy
-    ? "This session is running somewhere else right now — resuming it would interleave both transcripts"
-    : `Restore this session · last active ${when}`;
-  return `<div class="srow pastrow${busy ? " busy" : ""} ${d.id === pastMirrorId() ? "active" : ""}" data-past="${d.id}" data-key="${esc(d.colorKey)}" title="${esc(tip)}">
-    <span class="sglyph g-ended">·</span>
-    <span class="sbranch">${esc(label)}</span>
-    <span class="past-tag">${busy ? "busy" : when}</span>
-    <span class="sclose" data-forget="${d.id}" title="Remove from list — the conversation stays on disk">✕</span></div>`;
-}
-// A session that's live right now must not be offered for restore: Claude doesn't
-// lock the transcript, so a second --resume of the same id silently interleaves
-// both conversations into one file.
-function dormantBusy(d: Restorable): boolean {
-  for (const s of sessions.values()) if (s.resumeId === d.resumeId || s.id === d.id) return true;
-  return externals.some((e) => e.session_id === d.resumeId);
-}
-function extRow(e: ExtSession, chip?: WtCluster): string {
-  const working = extWorking(e);
-  const chipHtml = chip
-    ? `<span class="chip" style="--wtc:${branchHue(chip)}"><span class="fork">⑃</span><span class="lbl">${esc(chip.branch)}</span></span>`
-    : "";
-  return `<div class="srow extrow${chip ? " o3e" : ""} ${e.session_id === extMirrorId() ? "active" : ""}" data-ext="${e.session_id}" data-key="${esc(e.cwd)}">
-    <span class="sglyph ${working ? "g-work" : "g-idle"}">${working ? "●" : "○"}</span>
-    <span class="sbranch">${esc(e.name || basename(e.cwd))}</span>${chipHtml}
-    <span class="ext-tag" title="Running outside Episko · Claude v${esc(e.version)} · pid ${e.pid}">ext</span>
-    <span class="sjump" data-jump="${e.pid}" title="Jump to its terminal ↗">↗</span></div>`;
-}
+// The sidebar's row builders now live in ./sidebarview; renderSidebar below owns
+// the element they are painted into, and the drag state that must not be stomped.
 function renderSidebar() {
   // Don't stomp the DOM the browser is mid-drag on — see draggingProjects.
   if (draggingProjects) return;
@@ -1256,15 +1157,6 @@ async function maybeRunOnStop(s: Sess) {
   }
 }
 
-// The trailing column in the sidebar, and the palette subtitle. A background run
-// never claims to be finished, so it reads "bg" for as long as it lives.
-function taskStateText(s: Sess): string {
-  const r = s.run;
-  if (!r) return "";
-  if (s.phase === "working") return r.background ? "bg" : fmtShort(Date.now() - r.startedAt);
-  if (r.exitCode == null) return "";
-  return r.exitCode === 0 ? fmtShort(Date.now() - r.startedAt) : `exit ${r.exitCode}`;
-}
 // "a, b and c" — the quit guard lists up to three kinds of running thing.
 function listPhrase(parts: string[]): string {
   if (parts.length <= 1) return parts[0] ?? "";
